chore: move codeGenerator into folder
