+++ FABulous/FABulous_API.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from loguru import logger
 
 import FABulous.fabric_cad.model_generation_npnr as model_gen_npnr
-import FABulous.fabric_generator.code_generator as codeGen
+import FABulous.fabric_generator.code_generator.code_generator as codeGen
 import FABulous.fabric_generator.parser.parse_csv as fileParser
 from FABulous.fabric_cad.gen_bitstream_spec import generateBitsStreamSpec
 from FABulous.fabric_cad.gen_design_top_wrapper import generateUserDesignTopWrapper
@@ -14,7 +14,7 @@ from FABulous.fabric_definition.Bel import Bel
 from FABulous.fabric_definition.Fabric import Fabric
 from FABulous.fabric_definition.SuperTile import SuperTile
 from FABulous.fabric_definition.Tile import Tile
-from FABulous.fabric_generator.code_generation_VHDL import VHDLWriter
+from FABulous.fabric_generator.code_generator.code_generation_VHDL import VHDLWriter
 from FABulous.fabric_generator.fabric_automation import genIOBel
 from FABulous.fabric_generator.gen_fabric.gen_configmem import generateConfigMem
 from FABulous.fabric_generator.gen_fabric.gen_fabric import generateFabric
