+++ FABulous/fabric_generator/fabric_automation.py
@@ -11,13 +11,15 @@ from FABulous.fabric_definition.Bel import Bel
 from FABulous.fabric_definition.define import IO, MultiplexerStyle
 from FABulous.fabric_definition.Gen_IO import Gen_IO
 from FABulous.fabric_definition.Port import Port
-from FABulous.fabric_generator.code_generation_Verilog import VerilogWriter
-from FABulous.fabric_generator.code_generation_VHDL import VHDLWriter
+from FABulous.fabric_generator.code_generator.code_generation_Verilog import (
+    VerilogWriter,
+)
+from FABulous.fabric_generator.code_generator.code_generation_VHDL import VHDLWriter
 from FABulous.fabric_generator.parser.parse_hdl import parseBelFile
 from FABulous.fabric_generator.parser.parse_switchmatrix import parseList
 
 if TYPE_CHECKING:
-    from FABulous.fabric_generator.code_generator import codeGenerator
+    from FABulous.fabric_generator.code_generator.code_generator import codeGenerator
 
 
 def generateCustomTileConfig(tile_path: Path) -> Path:
