+++ FABulous/fabric_generator/gen_fabric/gen_fabric.py
@@ -2,8 +2,8 @@ from pathlib import Path
 
 from FABulous.fabric_definition.define import IO, ConfigBitMode, Direction
 from FABulous.fabric_definition.Fabric import Fabric
-from FABulous.fabric_generator.code_generation_VHDL import VHDLWriter
-from FABulous.fabric_generator.code_generator import codeGenerator
+from FABulous.fabric_generator.code_generator.code_generation_VHDL import VHDLWriter
+from FABulous.fabric_generator.code_generator.code_generator import codeGenerator
 
 
 def generateFabric(writer: codeGenerator, fabric: Fabric) -> None:
