+++ FABulous/fabric_generator/gen_fabric/gen_tile.py
@@ -6,9 +6,11 @@ from FABulous.fabric_definition.define import IO, ConfigBitMode, Direction
 from FABulous.fabric_definition.Fabric import Fabric
 from FABulous.fabric_definition.SuperTile import SuperTile
 from FABulous.fabric_definition.Tile import Tile
-from FABulous.fabric_generator.code_generation_Verilog import VerilogWriter
-from FABulous.fabric_generator.code_generation_VHDL import VHDLWriter
-from FABulous.fabric_generator.code_generator import codeGenerator
+from FABulous.fabric_generator.code_generator.code_generation_Verilog import (
+    VerilogWriter,
+)
+from FABulous.fabric_generator.code_generator.code_generation_VHDL import VHDLWriter
+from FABulous.fabric_generator.code_generator.code_generator import codeGenerator
 
 
 def generateTile(writer: codeGenerator, fabric: Fabric, tile: Tile) -> None:
