+++ FABulous/fabric_generator/gen_fabric/gen_top_wrapper.py
@@ -3,9 +3,11 @@ from pathlib import Path
 
 from FABulous.fabric_definition.define import IO
 from FABulous.fabric_definition.Fabric import Fabric
-from FABulous.fabric_generator.code_generation_Verilog import VerilogWriter
-from FABulous.fabric_generator.code_generation_VHDL import VHDLWriter
-from FABulous.fabric_generator.code_generator import codeGenerator
+from FABulous.fabric_generator.code_generator.code_generation_Verilog import (
+    VerilogWriter,
+)
+from FABulous.fabric_generator.code_generator.code_generation_VHDL import VHDLWriter
+from FABulous.fabric_generator.code_generator.code_generator import codeGenerator
 
 
 def generateTopWrapper(writer: codeGenerator, fabric: Fabric) -> None:
